# Improve perf test script and add script for comparing benchmarks