--- a/tests/perf/grpc-perf/etcd.conf
+++ b/tests/perf/grpc-perf/etcd.conf
@@ -1,4 +1,0 @@
-insecure-transport: true
-dial-timeout: 1000000000
-endpoints:
- - "172.17.0.1:2379"