This PR adds script for comparing benchmarks. By default benchmarks package `plugins/kvscheduler`, but can be used for any package, by using argument for the script.

Example output:
```
$ ./scripts/cmpbench.sh
-> running cmpbench for package: github.com/ligato/vpp-agent/plugins/kvscheduler
-> found old bench at: /tmp/bench/old.txt
--------------------------------------------------------------
goos: linux
goarch: amd64
pkg: github.com/ligato/vpp-agent/plugins/kvscheduler
BenchmarkScale/1-4 	   10000	    335051 ns/op	   47821 B/op	     768 allocs/op
BenchmarkScale/10-4         	    2000	   1539483 ns/op	  255235 B/op	    3738 allocs/op
BenchmarkScale/100-4        	     300	  17223630 ns/op	 2337862 B/op	   32479 allocs/op
BenchmarkScale/1000-4       	       5	 758695487 ns/op	24663652 B/op	  332955 allocs/op
PASS
ok  	github.com/ligato/vpp-agent/plugins/kvscheduler	18.223s
--------------------------------------------------------------
-> creating new bench..
--------------------------------------------------------------
goos: linux
goarch: amd64
pkg: github.com/ligato/vpp-agent/plugins/kvscheduler
BenchmarkScale/1-4 	   10000	    335938 ns/op	   47821 B/op	     769 allocs/op
BenchmarkScale/10-4         	    2000	   1581790 ns/op	  255236 B/op	    3737 allocs/op
BenchmarkScale/100-4        	     200	  17353561 ns/op	 2337766 B/op	   32479 allocs/op
BenchmarkScale/1000-4       	       5	 833877061 ns/op	24667713 B/op	  332986 allocs/op
PASS
ok  	github.com/ligato/vpp-agent/plugins/kvscheduler	17.018s
--------------------------------------------------------------
-> new bench at: /tmp/bench/old.txt
-> comparing benchmarks..
--------------------------------------------------------------
benchmark                 old ns/op     new ns/op     delta
BenchmarkScale/1-4        335051        335938        +0.26%
BenchmarkScale/10-4       1539483       1581790       +2.75%
BenchmarkScale/100-4      17223630      17353561      +0.75%
BenchmarkScale/1000-4     758695487     833877061     +9.91%

benchmark                 old allocs     new allocs     delta
BenchmarkScale/1-4        768            769            +0.13%
BenchmarkScale/10-4       3738           3737           -0.03%
BenchmarkScale/100-4      32479          32479          +0.00%
BenchmarkScale/1000-4     332955         332986         +0.01%

benchmark                 old bytes     new bytes     delta
BenchmarkScale/1-4        47821         47821         +0.00%
BenchmarkScale/10-4       255235        255236        +0.00%
BenchmarkScale/100-4      2337862       2337766       -0.00%
BenchmarkScale/1000-4     24663652      24667713      +0.02%
--------------------------------------------------------------
-> bench comparison image at: /tmp/bench/cmp.svg
```

Example for generated benchmark comparison image:

![cmp](https://user-images.githubusercontent.com/32484950/54417771-69fa5c00-4703-11e9-9ecb-f91a8c1970ca.png)